Staging

+++ app/Http/Controllers/Admin/AffilateController.php
@@ -11,6 +11,8 @@ use App\Models\Category;
 use App\Models\SubCategory;
 use App\Models\ChildCategory;
 use App\Models\AffiliateProduct;
+use Illuminate\Support\Facades\Redirect;
+use Validator;
 use Config;
 
 class AffilateController extends Controller
@@ -27,7 +29,7 @@ class AffilateController extends Controller
 
     public  function  index()
     {
-        $affiliateProducts = AffiliateProduct::all();
+        $affiliateProducts = AffiliateProduct::orderBy('created_at', 'DESC')->get();
         $status = Config::get('constants.status');
         $type = Config::get('constants.type');
         return view('admin.affilateproduct.list', compact('affiliateProducts', 'type','status'));
@@ -36,15 +38,40 @@ class AffilateController extends Controller
     public  function  create()
     {
         $categories = Category::all();
-        return view('admin.affilateproduct.create', compact('categories'));
+        $conditions = Config::get('constants.condition');
+        return view('admin.affilateproduct.create', compact('categories', 'conditions'));
     } 
     //save affiliated products...
     public  function  store(Request $request)
     {
+        //--- Validation Section
+        $rules = [
+             'a_name' => 'required',
+             'avatar' => 'required|image|mimes:jpeg,png,jpg,gif,svg|max:2048',
+             'a_sku' => 'required|unique:affiliate_products|',
+             'con' => 'required',
+             'a_ship_time' => 'required',
+             'a_color' => 'required',
+             'a_catog' => 'required',
+             'a_sub_catog' => 'required',
+             'a_new_price' => 'required',
+             'a_old_price' => 'required',
+             'a_stock' => 'required',
+
+        ];
+        $customs = [
+             'avatar.mimes' => 'File must be with extension jpeg,png,jpg,gif or svg.',
+             'avatar.max' => 'Fil size must be less than 2MB.',
+             'a_sku.unique' => 'This SKU has already been taken.',
+             'a_sku.regex' => 'SKU Must Not Have Any Special Characters.',
+        ];
+        $validator = Validator::make($request->all(), $rules, $customs);
+
+        if($validator->fails()) {
+            return Redirect::back()->withErrors($validator);
+        }
+        //end validation
         $product = new AffiliateProduct();
-        $product->admin_id = Auth::guard('admin')->id();
-        $product->a_type = $request->input('p_type');
-        $product->a_name = $request->input('p_name');
         if($request->hasfile('avatar'))
         {
             $file = $request->file('avatar');
@@ -54,38 +81,42 @@ class AffilateController extends Controller
             $product->a_image = $filename;
 
         }
-        $product->a_sku = $request->input('p_sku');
-        $product->a_size_qnty = $request->input('p_size_qnty');
-        $product->a_size_price = $request->input('p_size_price');
-        $product->a_ship_time = $request->input('p_ship_time');
-        $product->a_color = $request->input('p_color');
-        $product->a_catog = $request->input('p_catog');
-        $product->a_sub_catog = $request->input('p_sub_catog');
-        $product->a_child_catog = $request->input('p_child_catog');
-        $product->a_new_price = $request->input('p_new_price');
-        $product->a_old_price = $request->input('p_old_price');
-        $product->a_stock = $request->input('p_stock');
-        $product->a_detail = $request->input('p_detail');
-        $product->a_r_policy = $request->input('p_r_policy');
-        if($request->input('small') == true){
+        $product->admin_id = Auth::guard('admin')->id();
+        $product->a_type = $request->input('a_type');
+        $product->a_name = $request->input('a_name');
+        $product->a_sku = $request->input('a_sku');
+        $product->plink = $request->input('plink');
+        if($request->input('url_link') == NULL){
+            $product->url_link = 'https://www.youtube.com/'; 
+        }else{
+            $product->url_link = $request->input('url_link');
+        }
+        $product->con = $request->input('con');
+        $product->a_ship_time = $request->input('a_ship_time');
+        $product->a_color = $request->input('a_color');
+        $product->a_catog = $request->input('a_catog');
+        $product->a_sub_catog = $request->input('a_sub_catog');
+        $product->a_child_catog = $request->input('a_child_catog');
+        $product->a_new_price = $request->input('a_new_price');
+        $product->a_old_price = $request->input('a_old_price');
+        $product->a_stock = $request->input('a_stock');
+        $product->a_detail = $request->input('a_detail');
+        $product->a_r_policy = $request->input('a_r_policy');
+        $product->a_status = 1;
+        if($request->input('a_small') == true){
             $product->a_small = 1;
         }else{
             $product->a_small = 0;
         }
-        if($request->input('medium') == true){
+        if($request->input('a_medium') == true){
             $product->a_medium = 1;
         }else{
             $product->a_medium = 0;
         }
-        if($request->input('large') == true){
+        if($request->input('a_large') == true){
             $product->a_large = 1;
         }else{
             $product->a_large = 0;
-        }
-        if($request->input('checkbox') == true){
-            $product->a_status = 1;
-        }else{
-            $product->a_status = 0;
         }
         $product->save();
         return redirect()->route('admin.affilateproduct_list');
@@ -95,28 +126,52 @@ class AffilateController extends Controller
 
     public  function  edit($id)
     {
-        
-        $product = AffiliateProduct::where('id', $id )->first();
+        $product = AffiliateProduct::where('affiliate_products.id', $id )
+                ->leftJoin('subcategories', 'a_sub_catog', '=', 'subcategories.id')
+                ->leftJoin('childcategories', 'a_child_catog', '=', 'childcategories.id')
+                ->select('affiliate_products.*', 'subcategories.name as sub_name', 'childcategories.name as child_name')
+                ->first();
         $categories = Category::all();
-        $catog = Category::where('id', $product->a_catog)->first();
-        $sub_cat = SubCategory::where('id', $product->a_sub_catog)->first();
-        $ch_cat = ChildCategory::where('id', $product->a_child_catog)->first();
         $type = Config::get('constants.type');
-        return view('admin.affilateproduct.edit', compact('product', 'categories', 'type', 'catog','sub_cat','ch_cat'));
+        $conditions = Config::get('constants.condition');
+        return view('admin.affilateproduct.edit', compact('product', 'categories', 'type', 'conditions'));
         
     }
 
         // update ...
         public function update(Request $request, $id)
         {
+        //--- Validation Section
+        $rules = [
+            'a_name' => 'required',
+            'avatar' => 'image|mimes:jpeg,png,jpg,gif,svg|max:2048',
+            'a_sku' => 'required',
+            'con' => 'required',
+            'a_ship_time' => 'required',
+            'a_color' => 'required',
+            'a_catog' => 'required',
+            'a_sub_catog' => 'required',
+            'a_new_price' => 'required',
+            'a_old_price' => 'required',
+            'a_stock' => 'required',
+       ];
+       $customs = [
+            'avatar.mimes' => 'File must be with extension jpeg,png,jpg,gif or svg.',
+            'avatar.max' => 'Fil size must be less than 2MB.',
+            'a_sku.regex' => 'SKU Must Not Have Any Special Characters.',
+
+       ];
+       $validator = Validator::make($request->all(), $rules, $customs);
+
+       if($validator->fails()) {
+           return Redirect::back()->withErrors($validator);
+       }
+       //end validation
             $product = AffiliateProduct::where('id', $id)->first();
-    
-            $product->a_type = $request->input('p_type');
-            $product->a_name = $request->input('p_name');
             if($request->hasfile('avatar'))
             {
                 //new 
-                $destination = 'storage/uploads/products/'.$product->p_image;
+                $destination = 'storage/uploads/products/'.$product->a_image;
                 if(File::exists($destination))
                     {
                         File::delete($destination);
@@ -129,38 +184,44 @@ class AffilateController extends Controller
                 $product->a_image = $filename;
     
             }
-            $product->a_sku = $request->input('p_sku');
-            $product->a_size_qnty = $request->input('p_size_qnty');
-            $product->a_size_price = $request->input('p_size_price');
-            $product->a_ship_time = $request->input('p_ship_time');
-            $product->a_color = $request->input('p_color');
-            $product->a_catog = $request->input('p_catog');
-            $product->a_sub_catog = $request->input('p_sub_catog');
-            $product->a_child_catog = $request->input('p_child_catog');
-            $product->a_new_price = $request->input('p_new_price');
-            $product->a_old_price = $request->input('p_old_price');
-            $product->a_stock = $request->input('p_stock');
-            $product->a_detail = $request->input('p_detail');
-            $product->a_r_policy = $request->input('p_r_policy');
-            if($request->input('small') == true){
+            $product->a_name = $request->input('a_name');
+            $product->a_sku = $request->input('a_sku');
+            $product->plink = $request->input('plink');
+            if($request->input('url_link') == NULL){
+                $product->url_link = 'https://www.youtube.com/'; 
+            }else{
+                $product->url_link = $request->input('url_link');
+            }
+            $product->con = $request->input('con');
+            $product->a_ship_time = $request->input('a_ship_time');
+            $product->a_color = $request->input('a_color');
+            $product->a_catog = $request->input('a_catog');
+            $product->a_sub_catog = $request->input('a_sub_catog');
+            $product->a_child_catog = $request->input('a_child_catog');
+            $product->a_new_price = $request->input('a_new_price');
+            $product->a_old_price = $request->input('a_old_price');
+            $product->a_stock = $request->input('a_stock');
+            $product->a_detail = $request->input('a_detail');
+            $product->a_r_policy = $request->input('a_r_policy');
+            if($request->input('a_small') == true){
                 $product->a_small = 1;
             }else{
                 $product->a_small = 0;
             }
-            if($request->input('medium') == true){
+            if($request->input('a_medium') == true){
                 $product->a_medium = 1;
             }else{
                 $product->a_medium = 0;
             }
-            if($request->input('large') == true){
+            if($request->input('a_large') == true){
                 $product->a_large = 1;
             }else{
                 $product->a_large = 0;
             }
             if($request->input('checkbox') == true){
-                $product->a_status = 1;
+                $product->status = 1;
             }else{
-                $product->a_status = 0;
+                $product->status = 0;
             }
             $product->update();
             return redirect()->route('admin.affilateproduct_list');
@@ -168,13 +229,13 @@ class AffilateController extends Controller
     
         public  function  view($id)
         {
-            $product = AffiliateProduct::where('id', $id)->where('vendor_id', Auth::guard('vendor')->id())->first();
+            $product = AffiliateProduct::where('id', $id)->first();
             return view('admin.affilateproduct.view', compact('product'));
         }
         public function destroy($id)
         {
             $product = AffiliateProduct::where('id', $id)->first();
-            $destination = 'storage/uploads/products/'.$product->p_image;
+            $destination = 'storage/uploads/products/'.$product->a_image;
             if(File::exists($destination))
               {
                   File::delete($destination);
